test: integration test for both v0.1 and v0.2 getStateUpdate
