+++ crates/pathfinder/src/rpc/v02/method.rs
@@ -13,7 +13,7 @@ pub(super) mod get_class_at;
 pub(super) mod get_class_hash_at;
 pub(super) mod get_events;
 pub(super) mod get_nonce;
-pub(super) mod get_state_update;
+pub(crate) mod get_state_update;
 pub(super) mod get_storage_at;
 pub(super) mod get_transaction_by_block_id_and_index;
 pub(super) mod get_transaction_by_hash;
